[QFEED-141] Refactor: 추천 애그리거트를 팔로우에서 분리
